Transfrom_Load_to_Bigquery/requirements.txt
functions-framework
google-cloud-storage
google-cloud-bigquery
flask
